+++ src/ansys/fluent/core/services/reduction.py
@@ -29,7 +29,7 @@ import grpc
 
 from ansys.api.fluent.v0 import reduction_pb2 as ReductionProtoModule
 from ansys.api.fluent.v0 import reduction_pb2_grpc as ReductionGrpcModule
-from ansys.fluent.core.physicalquantities.strategies.fluent import FluentExprStrategy
+from ansys.fluent.core.quantity_strategies import FluentExprStrategy
 from ansys.fluent.core.services.datamodel_se import _convert_variant_to_value
 from ansys.fluent.core.services.interceptors import (
     BatchInterceptor,
