+++ src/ansys/fluent/core/services/solution_variables.py
@@ -32,8 +32,8 @@ import numpy as np
 from ansys.api.fluent.v0 import field_data_pb2 as FieldDataProtoModule
 from ansys.api.fluent.v0 import svar_pb2 as SvarProtoModule
 from ansys.api.fluent.v0 import svar_pb2_grpc as SvarGrpcModule
-from ansys.fluent.core.physicalquantities.strategies.fluent import FluentSVarStrategy
 from ansys.fluent.core.pyfluent_warnings import PyFluentDeprecationWarning
+from ansys.fluent.core.quantity_strategies import FluentSVarStrategy
 from ansys.fluent.core.services.field_data import (
     _FieldDataConstants,
     override_help_text,
